+++ NEWS.md
@@ -1,5 +1,7 @@
 # version 0.6-4
 
+* `plot.stars()` and `image.stars()` are sensitive to `cex.axis`, for axes and key (requires sf >= 1.0-14); #642
+
 * move `lwgeom` dependency to Suggests; using `st_transform_proj()` requires loading `lwgeom` first
 
 # version 0.6-3